requirements.txt
pyTelegramBotAPI==4.20.0
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.5
